backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pypdf
python-docx
pydantic
python-dotenv
